Support changing the dev mode image for `hack/generate-manifest.sh`
This patch adds the support for specifying the image used for deploying
the antrea pods in case of dev mode. This is needed for debugability
in case a custom image is required to be tested with the latest manifests
or in case the image was built directly using `docker build` and the tag
is changed.