[CI] generate antrea manifests in hw-offload CI with a custom image
Following #1617, the default image name in the antrea.yml file changed
to use the harbor image. Although the patch includes a change in the
make file to change the image name correctly in case the image was
built, the hw-offload CI uses `docker build` to build the image, and
this made it so that the CI would not use the image it was building,
but the upstream latest image.

Also changed the way to enable the hw-offload to use the generate-manifests.sh
instead of using `sed`.

This patch fixes that by building the antrea manifests with the image
name being the same as the one that was built.