Add script for semi-automatic bug fix backporting (#2136)
Just like the Kubernetes project, I think we should start asking
contributors to backport their critical bug fixes themselves when
required. The main benefit is that contributors are in a better position
to resolve possible merge conflicts than whoever is responsible for the
patch release. If the contributor cannot take care of it, a maintainer
can still step up and handle that task.

To simplify this process, we provide a script which will take care of
cherry-picking the change to the correct release branch, and will
automatically open a PR on Github, with the 'kind/cherry-pick'
label. The script is basically a copy of the one provided for
Kubernetes.

We also add documentation to guide contributors through the
cherry-picking process.

Fixes #1532