Add memberlist cluster in antrea-agent for egress failover support
1.Add memberlist cluster in antrea-agent for egress failover support

A cluster will be created in the background when the egress feature is turned on.
When the local Node's labels match the nodeSelector from at least one ExternalIPPool,
the Node will join then cluster. Similarly, the Node will leave cluster if its labels are
updated and no longer match the nodeSelector from any ExternalIPPool.

Each Node in the cluster holds the same consistent hash ring for each ExternalIPPool,
in order to distribute egress IPs equally among the selected Nodes (which are part of the
memberlist cluster). When a Node leaves the cluster, its IPs are redistributed.
When a Node joins the cluster, it's added to the hash ring and a small fraction of IPs are re-assigned to that Node.

2.Add selecting egress's owner node and assign egress-ip to node

Assign a owner node for egress which with a valid externalIPPool.
Add egress status of api, when egress has assigned a owner node and egressIP has assigned,
the egress status will updated, egress status is the owner node name.

Signed-off-by: wenqiq <wenqiq@vmware.com>