Rename Go module from github.com/vmware-tanzu/antrea to antrea.io/antrea (#2166)
This is a pretty straightforward change with one exception: we cannot
use the default Protobuf package names for Antrea APIs because of name
resolution in protoc and because of how go-to-protobuf references
messages defined in other .proto files. go-to-protobuf will not add a
leading dot to the message path, and in that case protoc does resolution
starting from the inbermost scope. Default package names start with
"antrea.io.antrea." which means that for any message reference that
matches "antrea.io.antrea.<path>", protoc will look for
"antrea.io.antrea.io.antrea.<path>" which obviously will fail. To avoid
a "sed" hack to add a leading dot (which would require us to run protoc
directly), we rename the package names to "antrea_io.antrea.*". Even
though this causes some extra code to be re-generated, there shouldn't
be any impact at all on backwards-compatibility: the Protobuf messages
themselves do not change.

Consumers of the Antrea Go module should update the import path they use
for Antrea. If these consumers are "pinning" a version of Antrea which
is on a release branch that was created before the repository transfer
(for now, all release branches...), their code should keep working.

See #2154