Improve readability of NeedsRoutingToPeer method (#2252)
It was confusing that antrea-agent needed to add a route to the routing
table when NeedsRoutingToPeer returns false. "Need Routing" here meant
the Node cannot reach peer Node directly and it needs routing support
from the network connecting the Nodes. To make it clear, this patch
changes the method name to "NeedsDirectRoutingToPeer" and returns true
if the Node can reach peer Node directly which indicates antrea-agent
needs to add a direct route to the routing table.

Signed-off-by: Quan Tian <qtian@vmware.com>