Protect Traceflow state from concurrent access in Octant plugin (#2261)
After checking with the Octant developers, we need to protect the global
plugin state in case multiple client sessions are opened (e.g. multiple
tabs). While this is not a typical use case, it could potentially lead
to some data corruption.

We don't try to get fancy with the locking scheme, i.e. we don't try to
keep the size of the critical section small, as it doesn't seem to
matter much in this case. The goal is to ensure code correctness.

Note that the K8s client itself is thread-safe, and it is never updated
after initialiation.

Signed-off-by: Antonin Bas <abas@vmware.com>