Add e2e tests for Pod-to-External flow type (#2231)
In the current flow-aggregator e2e tests, we have tests for flow types: intra-node and inter-node. This patch adds a test case for pod-to-external flow type. It takes a busybox pod as the source pod (client), a hostNetwork nginx pod as the destination (server). Server's IP will be the node's IP where it is deployed on. Let the client send requests to the server, then we generate pod-to-external flows. Finally we check the exports of IPFIX flow records.

Signed-off-by: heanlan <anlan9771@gmail.com>