Fix race condition in EgressController and unit tests
ipAllocatorMap was read without holding the mutex in syncEgressIP. This
patch fixes it and stabilizes the unit tests by waiting for cache sync
before starting creating objects. This is because resource creation
events will be missing if the resources are created in-between list and
watch call of an fake informer.

Signed-off-by: Quan Tian <qtian@vmware.com>