[Windows]Check HostNamespace value when removing HcnEndpoint (#2306)
For docker on Windows, no HostComputeNamespace is created before calling
CNI to add network configuration. But containerd would do it. If a
HnsEndpoint is created without HostComputeNamespace, a default value
"00000000-0000-0000-0000-000000000000" should be set. HCN API should
return an Error if a HnsEndpoint is removed from host Namespace
"00000000-0000-0000-0000-000000000000". Hence, we skip invocation of
"hcn.RemoveNamespaceEndpoint" for docker case.

Signed-off-by: wenyingd <wenyingd@vmware.com>