Add Namespaced Group CRD for Antrea NetworkPolicy
Add Group CRD which is responsible for collecting Pods and Namespaces in the Group's Namespace based on
labelselectors defined in the Group definition. It also allows setting an IPBlock and ChildGroups (cannot be set with other
Selectors) in the Group. The purpose of a Group is to allow grouping of resources and then be referenced in
AntreaNetworkPolicies without having to add the same selectors in every ANP when the group of resources are meant to be
shared. This allows for greater sharing and decouples the job of reconciling effective group members from that of enforcing
security policies.

This PR adds the following:
-Group API types
-Group CRD YAML
-Controller changes to reconcile effective members of a Group
-Controller changes to trigger ANP update introduced by a Group
-Validation webhook to validate a GroupSpec
-NetworkPolicyStatus refactored with Conditions

Signed-off-by: Qiyue Yao <yaoq@vmware.com>
Co-authored-by: abhiraut <rauta@vmware.com>