Fix flakiness in Kind e2e flow aggregator tests (#2308)
- Flows are not getting expired as expected because of zero start time and
last export time is initialized as start time.This commit fixes that bug by
adding current time as start time if it is zero when a new connection is added.

- Idle flow export timeout is not configured properly in the e2e test.
There is a bug in the test and this commit fixes that.

- The expected number of records is not correct in Kind clusters
as packet counters are not supported in conntrack entries for
OVS userspace datapath. Flow exporter cannot send records after
active timeout expiry as we cannot consider flows to be active
without the statistics. Therefore, we expect only 2 records for any
given flow in Kind clusters and 3 flow records in other clusters.
We expect this limitation to be resolved soon when OVS
userspace datapath support statistics in conntrack entries.

- Additionally, we consider source port from iperf flows to ignore other
similar flow records from the iperf control flows. Because of this we
were passing the tests in Kind clusters even though there are only two flow
records.

- Kubectl logs command with since option is not outputting the logs of
ipfix collector as expected and this causes intermittent failures. Removing
the since option as we have source port as the gating parameter when processing
flow records.

Signed-off-by: Srikar Tati <stati@vmware.com>