Graduate NetworkPolicyStats feature from Alpha to Beta (#2351)
The feature has been around for many releases. We are currently testing
it on a large cluster, with a large number of NetworkPolicies, with no
issue so far.

On the e2e test side, we follow the same logic as for AntreaPolicy and
other Beta features: if the feature has been disabled, we skip all the
relevant tests that depend on that feature instead of updating the
ConfigMap as part of the test (which is the approach we take for Alpha
features). This commit cleans up the code we use to skip tests if a
feature is disabled. Note that there was a typo in
skipIfTraceflowDisabled: the function was checking for the AntreaProxy
feature instead of Traceflow.

Fixes #2199

Signed-off-by: Antonin Bas <abas@vmware.com>