Add FQDN policy feature in Antrea-native policies
This PR adds an `fqdn` field to the egress rules of Antrea-native
policy. It can be used to restrict egress access to the Fully Qualified
Domain Names, specified either by exact name or wildcard matchPatterns.

Supported formats for the `fqdn` field are:

* Exact FQDNs, i.e. `facebook.com`, `maps.google.com` or
  `db-svc.default.svc.cluster.local`.
* Wildcard expressions, i.e. `*wayfair.com` or `*.edu`.

The standard `Allow`, `Drop` and `Reject` actions apply to FQDN egress
rules. In a single NetworkPolicyPeer, `fqdn` should not be specified
together with pod/namespaceSelectors, ClusterGroups or IPBlocks.

Note that this feature is not a L7 policy and only works at L3/L4. DNS
packets are intercepted and inspected only to help the fqdn controller
cache resolved addresses. DNS response packets will only be dropped if
the client's requested FQDN is matched by a FQDN policy rule, but the
datapath reconciliation for that rule fails.

Co-authored-by: Antonin Bas <abas@vmware.com>
Co-authored-by: Grayson Wu <wgrayson@vmware.com>

Signed-off-by: Antonin Bas <abas@vmware.com>