Support EgressIP assigning and failover in antrea-agent (#2186)
1.Add memberlist cluster in antrea-agent

A cluster will be created in the background when the Egress feature is
turned on. And the local Node will join all the other K8s Nodes in a
memberlist cluster.

Each Node in the cluster holds the same consistent hash ring for each
ExternalIPPool, in order to distribute egress IPs equally among the
selected Nodes (which are part of the memberlist cluster). When a Node
leaves the cluster, its IPs are redistributed. When a Node joins the
cluster, it's added to the hash ring and a small fraction of IPs are
re-assigned to that Node.

2.Assign EgressIP to owner Node and update Egress status

Assign a owner node for Egress which with a valid externalIPPool. When
Egress has been assigned a owner node and egressIP has assigned, the
Egress status will updated, Egress status is the name of the Node that
holds the Egress IP.

Signed-off-by: wenqiq <wenqiq@vmware.com>