Assign egress IPs to dummy device instead of external device (#2345)
Instead of assigning Egress IPs to the device that is connected to the
external network and relying on persisting assigned IPs to files, this
patch creates a dummy device and uses it to hold the external IPs
configured by antrea-agent. From function's perspective, it's totally
same regardless of the device the IPs are assigned to. The advantages of
using a dummy device are as below:

1. It doesn't need to persist assigned IPs to files to know which IPs
were configured by antrea-agent.

2. It avoids touching user or system managed network device, and is easy
to exclude the device from being managed by network manager tools.

3. It is more friendly for troubleshooting.

Besides, this patch fixes an issue that old Egress IPs were not
unassigned from the Node when updating Egress's EgressIP, and makes
IPAssigner more generic so that it can be used by other features that
need assign IPs to the system.

It also adds e2e tests that covers basic workflow of CRUD of Egress and
failover scenario.

Signed-off-by: Quan Tian <qtian@vmware.com>