Change default port range for NodePortLocal
From 40000-41000 to 61000-62000, to avoid conflict with the default
ip_local_port_range on Linux. This may be a slighlty disruptive change
to users who rely on the NPL feature: NPL port mappings for all existing
Pods will be updated to use the new range. However, I consider this
acceptable and it's better to make this change while NPL is still an
Alpha feature. Note that this behavior is verified by an e2e test:
TestNPLChangePortRangeAgentRestart.

See #2381

Signed-off-by: Antonin Bas <abas@vmware.com>