Remove CHANGELOG update step from contribution instructions
The initial idea was that contributors would add an entry to the
"Unreleased" section of the CHANGELOG for every significant change
contributed to the project. This would have simplified release
management. However, this is an error-prone process and we never
actually implemented it. Instead we should look into auto-generating the
CHANGELOG for each release based on PR descriptions (see #2411).

Signed-off-by: Antonin Bas <abas@vmware.com>