Fix the deadlock between exporter and conntrack polling go routines
Deadlock is due to the access of connection map from exporter goroutine waiting
to acquire connection map lock to update the "DoneExport" flag of the stored
connection. At the same time, the connection polling goroutine acquires the
connection map lock waiting to acquire flow record map, which is acquired by the
exporter goroutine.
This was caught in scale testing.
Resolved this through a temporary fix by adding the same flag in the flow record data struct.

The connection and record deletion logic will be re-evaluated through PR #2360 as it refactors the related code quite a bit.

Signed-off-by: Srikar Tati <stati@vmware.com>