Add support for UDP in NodePortLocal
This allows the NPL controller in antrea-agent to program iptable
rules based on NPL enabled services' protocols.
This commit also includes a change in the Pod's NPL annotation schema,
which now contains a list of protocols, that bind to the nodeport.
e2e test added to check for UDP based services, pod annotation and
iptable checks, and traffic testing.
issue #2395

Signed-off-by: Shubham Chauhan <shchauhan@vmware.com>