Fix deadlock in initializing GroupEntityIndex with many groups (#2404)
When initializing GroupEntityIndex, if there were more than 1000 events
need to send, adding more Pods or ExternalEntity would be pending with
the lock acquired because the channel was full and the Run method that
consumed the channel didn't start. The Run method ran after the
setInitialCounts method, which also needed to acquire the lock, hence
deadlock.

This patch fixes it by starting GroupEntityIndex's Run method
individually. To avoid similar issues in the future, it also removes
the need of acquiring the lock in setInitialCounts and HasSync methods.

There is a known bug in sync/atomic when using 64-bit atomics on a
32-bit architecture (e.g. ARM). As a result precautions must be taken to
guarantee correct alignment of the uint64 value.

Signed-off-by: Quan Tian <qtian@vmware.com>
Signed-off-by: Antonin Bas <abas@vmware.com>
Co-authored-by: Antonin Bas <abas@vmware.com>