Update some Go module dependencies (#2452)
* k8s libraries: v0.18.4 => v0.18.19
* gogo/protobuf: v1.3.1 => v1.3.2

The main motivation is that gogo/protobuf v1.3.1 is affected by
CVE-2021-3121. gogo/protobuf v1.3.1 is still referenced in the go.sum because of
the following indirect dependencies:

github.com/Microsoft/hcsshim@v0.8.9
github.com/containerd/cgroups@v0.0.0-20200531161412-0dbf7f05ba59
k8s.io/apimachinery@v0.18.4

The first 2 are because som dependencies are importing an older version of
containerd, and this is orthogonal to Antrea. The last one is through go-ipfix,
and once again it should not affect Antrea.

Signed-off-by: Antonin Bas <abas@vmware.com>