Install all Endpoint flows belong to a Service via single bundle (#2520)
It was observed that installing a bundle took around 1~3ms. In
InstallEndpointFlows, agent created a bundle for each Endpoint and
installed them sequentially. If there are many Services and Endpoints,
it could take long time to install flows in this fashion when agent
restarts. For instance, it took ~78s to install all flows for 4K
Services and 28K Endpoints.

This patch uses single bundle for all Endpoint flows belong to a Service
to reduce the number of bundles. In the above scale, the execution time
was reduced from ~78s to ~30s.

Signed-off-by: Quan Tian <qtian@vmware.com>