Log maintenance bug fix in the Flow Aggregator
Periodic deletion of log files is missing in the Flow Aggreegator.
This logic periodically deletes extra files by making sure that max
number of files to be 4. This will ensure that no disk space is wasted
on the Kubernetes node running the Flow Aggregator.

This logic was missed initially when the logging in the Flow Aggregator is
added in the beginning adopting the same framework as the Antrea Agent and the
Antrea Controller.

Signed-off-by: Srikar Tati <stati@vmware.com>