Support --sort-by flag for "antctl get networkpolicy" for Agent
Previous to this change, the flag was only supported in "Controller
mode". This was a bit confusing when running "antctl get networkpolicy"
from the Agent Pod, because 1) the option shows in the examples from the
help message, and 2) the documentation doesn't make it clear that this
flag only works in "Controller mode".

Rather than update the documentation and add complexity in the help
message, I am adding support for that flag in "Agent mode". There
doesn't seem to be any reason not to.

Signed-off-by: Antonin Bas <abas@vmware.com>