Fix missing Flow Aggregator logs in the artifacts when tests failed
In this commit, we fix missing Flow Aggregator logs in the artifacts
when tests failed. Also we add an option of verbosity level of Flow
Aggregator when generating yamls like Antrea side.

Also add a warning log to help debugging the Pod labels e2e test failure
of Flow Aggregator.

Signed-off-by: Yongming Ding <dyongming@vmware.com>