Do not use the iptables NOTRACK target if not available
We currently use this target to install rules which skip tracking of
encapsulated traffic (VXLAN / Geneve). In some rare instances the kernel
may not have been built with CONFIG_NETFILTER_XT_TARGET_CT, and
trying to install these rules can cause iptables-restore to fail during
Agent intialization on Linux.

See #2635

This patch checks for NOTRACK support by looking for the xt_CT module
(can be built-in or loadable). This requires /lib/modules to be mounted
to the antrea-agent container (read-only).

I considered the following alternatives to this solution:

* call modprobe with --dry-run: it essentially the same thing, with
  modprobe looking at modules.dep (instead of listing files) and
  modules.builtin
* use https://github.com/u-root/u-root/tree/master/pkg/kmodule: again
  essentially the same thing, but we add another dependency
* copy modules.dep and modules.builtin to /var/run/antrea in the
  initContainer, and read them from the antrea-agent. This avoids having
  to mount /lib/modules to the antrea-agent container, but does add an
  extra pre-requisite to run the Agent.
* look at the error of iptables-restore and re-attempt without the
  NOTRACK target rules: very clunky in my opinion.

I'm open to switching to one of these alternatives or to yet another
solution I haven't considered.

If the target is not available, we simply skip installing the rules as
they are for performance only and the Agent can still work without them.

Signed-off-by: Antonin Bas <abas@vmware.com>