Fix Clean-AntreaNetwork.ps1
The script is currently failing with

```
At C:\k\antrea\Clean-AntreaNetwork.ps1:9 char:80
+ ... eter(Mandatory = $false)] [string] $OVSInstallDir = "C:\openvswitch",
+                                                                          ~
Missing expression after ','.
    + CategoryInfo          : ParserError: (:) [], ParentContainsErrorRecordException
    + FullyQualifiedErrorId : MissingExpressionAfterToken
```

I don't know if this is dependent on the PowerShell version, but it
seems that in our other scripts we don't end with a comma for the last
parameter in the list.

Thanks @mmottaghi for reporting this.

Signed-off-by: Antonin Bas <abas@vmware.com>