Add in-cluster svc fqdn e2e test
Add a new test case using in-cluster headless Services to test FQDN
policies, to avoid having a dependency on external connectivity. The reason we
use headless Service is that FQDN will use the IP from DNS A/AAAA records to
implement flows in the egress policy table. For a non-headless Service, the DNS
name resolves to the ClusterIP for the Service. But when traffic arrives to the
egress table, the dstIP has already been DNATed to the Endpoints IP by
AntreaProxy Service Load-Balancing, and the policies are not enforced correctly.
For a headless Service, the Endpoints IP will be directly returned by the DNS
server. In this case, FQDN based policies can be enforced successfully.

Signed-off-by: wgrayson <wgrayson@vmware.com>