Fix triggerCNPUpdates panic
When a ClusterGroup is created, updated, or deleted, `triggerCNPUpdates`
will be called to update ClusterNetworkPolicies that reference to it.
However, a ClusterNetworkPolicy may have not been processed by `addCNP`
when they are found in the informer cache, which means its internal
NetworkPolicy doesn't exist yet. `triggerCNPUpdates` would panic if it
assumes the internal NetworkPolicy always exist and uses the returned
value directly.

This patch fixes it by checking existence first before using the
returned internal NetworkPolicy.

Signed-off-by: Quan Tian <qtian@vmware.com>