Fix flaky TestClusterIPv4 e2e tests
The tests checking access from the Node / hostNetwork were broken on
Kind clusters: wget is used to check connectivity, but wget is not
installed on the Kind Node image. This issue was going unnoticed however
because the tests were not checking for errors properly.

The tests checking access from the Pod network were flaky (at least with
kube-proxy) because no more than one connection was attempted, with
little delay between Service creation and the connectivity
test. Instead, we replace wget with agnhost, and we make up to 5
connection attempts.

Fixes #2860

Signed-off-by: Antonin Bas <abas@vmware.com>