Improve multi-protocol support for NPL
There is an edge case, which is not handled well by the current NPL
controller implementation: if a given Pod port needs to be exposed for
both TCP and UDP, and the first available Node port is only available
for TCP, it will be still be selected and NPL rule installation will
succeed for TCP but not for UDP.

We have 2 options:

1. reserve the port for both protocols ahead of time, even if the port
is only needed for one protocol initially.
2. support using different Node ports for different protocols, even when
the Pod port is the same.

In this patch, we go with option 1) to preserve the "nice" property that
a give Pod port maps to a unique Node port.

Fixes #2894

Signed-off-by: Antonin Bas <abas@vmware.com>