Add a manual Github workflow to build and push antrea/ethtool (#2900)
Anyone with write permissions to antrea-io/antrea can run the workflow,
but they should check with a maintainer first. If someone runs the
workflow by mistake, it is very unlikely that anything will break.

The Docker image now supports ARM (arm64 and arm/v7) in addition to
amd64. This is useful for example if someone wants to run a Kind cluster
with Antrea on a MAC laptop with an M1 chip. See #2895.

Signed-off-by: Antonin Bas <abas@vmware.com>