Graduate NodePortLocal feature from Alpha to Beta
NodePortLocal will be graduated to Beta and enabled by default starting
with v1.4.

On the e2e test side, we follow the same logic as for other Beta
features: if the feature has been disabled, we skip all the relevant
tests that depend on that feature instead of updating the ConfigMap as
part of the test (which is the approach we take for Alpha
features). The NPL e2e tests can still update the ConfigMap if the
nplPortRange needs to be updated. This commit changes the
mutateAntreaConfigMap so that the function is a no-op in case there is
no actual change.

Fixes #2923

Signed-off-by: Antonin Bas <abas@vmware.com>