Move all config structs to their own package
Type definitions for the configuration structs for all Antrea components
(Agent, Controller, FlowAggregator) are moved to the pkg/config
package. This makes it easy to import the package in e2e tests, in order
to mutate the antrea-config ConfigMap when test cases require it.

I don't believe there is any downside to that change. I chose to group
all config structs in pkg/config instead of placing the definitions in
pkg/agent/config, pkg/controller/config, pkg/flowaggregator/config.

This simplifies the e2e test framework and makes the process of updating
the ConfigMap type-safe and less error-prone. The framework_test.go file
is no longer required.

As part of this change, I made a small change to the
FlowAggregatorConfig struct: the RecordContent field is now a struct and
no longer an arbitrary map. This should not impact users and makes the
struct more similar to AgentConfig and ControllerConfig.

Signed-off-by: Antonin Bas <abas@vmware.com>