Add startTime to the Traceflow Status
The startTime is used to determine if a Traceflow has timed out and
should be reported as failed.

Until now we were relying on the CreationTimestamp set by the
kube-apiserver. However, by chance, I was testing Traceflow on a cluster
with a clock skew between the control plane Node (running the
kube-apiserver) and the worker Node running the
antrea-controller. Because of the clock skew, each Traceflow request was
tagged as failed as soon as I created it (with reason timeout). It took
me a while to figure out the reason.

By introducing the startTime field (set & used by the
antrea-controller), we avoid the possibility of such issues. If
startTime is not available for any reason, we fall back to the
CreationTimestamp.

This API change is backwards-compatible.

Signed-off-by: Antonin Bas <abas@vmware.com>